Move logic for indexed document types inside config model api wrapping
